refactor: add editable components for event fields
