+++ src/components/Event/EventFields/Author.tsx
@@ -1,9 +1,8 @@
-import React, { type ReactNode } from 'react';
+import React from 'react';
 import clsx from 'clsx';
 import styles from './styles.module.scss';
 
 import { observer } from 'mobx-react-lite';
-import { useStore } from '@site/src/stores/hooks';
 import { ReadonlyProps } from './iEventField';
 import Badge from '@site/src/components/shared/Badge';
 
@@ -14,7 +13,7 @@ const Author = observer((props: ReadonlyProps) => {
             style={{ gridColumn: 'author' }}
             className={clsx('author', styles.author, props.className, 'grid-author')}
         >
-            <Badge text={(props.event.author?.shortName || props.event.author?.email) ?? '-'} />
+            <Badge text={(event.author?.shortName || event.author?.email) ?? '-'} />
         </div>
     );
 });
