+++ src/components/Event/EventFields/DescriptionLong/index.tsx
@@ -1,10 +1,10 @@
-import React, { type ReactNode } from 'react';
+import React from 'react';
 import clsx from 'clsx';
-import styles from './styles.module.scss';
+import styles from '../styles.module.scss';
 
 import { observer } from 'mobx-react-lite';
-import { Props as Base } from './iEventField';
-import TextArea from '@site/src/components/shared/TextArea';
+import { Props as Base } from '../iEventField';
+import Edit from './Edit';
 
 interface Props extends Base {
     displayMultiLine?: boolean;
@@ -13,17 +13,7 @@ interface Props extends Base {
 const DescriptionLong = observer((props: Props) => {
     const { event } = props;
     if (event.isEditable && event.isEditing) {
-        return (
-            <div
-                style={{ gridColumn: 'descriptionLong' }}
-                className={clsx(styles.descriptionLong, props.className, 'grid-descriptionLong')}
-            >
-                <TextArea
-                    text={event.descriptionLong}
-                    onChange={(text) => event.update({ descriptionLong: text })}
-                />
-            </div>
-        );
+        return <Edit {...props} />;
     }
     const displayMultiline = event.isExpanded || props.displayMultiLine;
     return (
