+++ src/components/Event/EventFields/Location/index.tsx
@@ -1,26 +1,16 @@
 import React, { type ReactNode } from 'react';
 import clsx from 'clsx';
-import styles from './styles.module.scss';
+import styles from '../styles.module.scss';
 
 import { observer } from 'mobx-react-lite';
-import { Props } from './iEventField';
+import { Props } from '../iEventField';
 import TextArea from '@site/src/components/shared/TextArea';
+import Edit from './Edit';
 
 const Location = observer((props: Props) => {
     const { event } = props;
     if (event.isEditable && event.isEditing) {
-        return (
-            <div
-                style={{ gridColumn: 'location' }}
-                className={clsx(styles.location, props.className, 'grid-Location')}
-            >
-                <TextArea
-                    text={event.location}
-                    onChange={(loc) => event.update({ location: loc })}
-                    rows={2}
-                />
-            </div>
-        );
+        return <Edit {...props} />;
     }
     return (
         <div
